+++ src/templates/config.ts
@@ -1,5 +1,3 @@
-export const configTemplate = (folderPath: string) => {
-      const _name = folderPath + "Config"
-      return `export const ${_name} = {
-}`
+export const configTemplate = () => {
+      return `export default {}`
 }