+++ src/templates/request.ts
@@ -1,8 +1,11 @@
 export const requestTemplate = (folderPath: string) => {
       const _name = folderPath + "Request"
-      return `export const ${_name} = {
-      index: () => {
-            console.log("index get request")
-      },
-}`
+      return `class ${_name} {
+      constructor() {}
+      index() {
+            console.log("hello world!")
+      }
+}
+export default new ${_name}()
+`
 }